co-ux-designer/requirements.txt
crewai>=0.80.0
crewai-tools>=0.20.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-anthropic>=0.3.0
openai>=1.0.0
